backend/requirements.txt
Flask==3.0.0
openai==1.30.1
PyPDF2==3.0.1
python-docx==0.8.11
email-validator==2.1.1
chromadb==0.4.15
python-dotenv==1.0.0
tiktoken==0.5.2
